+++ Source/Public/Get-VSTeamProcess.ps1
@@ -1,56 +1,40 @@
 function Get-VSTeamProcess {
    [CmdletBinding(DefaultParameterSetName = 'List')]
+  # [OutputType([vsteamprocess])]
    param(
-      [Parameter(ParameterSetName = 'List')]
+      [Parameter(ParameterSetName = 'ByName', Position=0)]
+      [ArgumentCompleter([ProcessTemplateCompleter])]
+      [Alias('ProcessName','ProcessTemplate')]
+      $Name = '*',
+
+      [Parameter(DontShow=$true, ParameterSetName = 'List')]
       [int] $Top = 100,
 
-      [Parameter(ParameterSetName = 'List')]
+      [Parameter(DontShow=$true, ParameterSetName = 'List')]
       [int] $Skip = 0,
 
       [Parameter(ParameterSetName = 'ByID')]
       [Alias('ProcessTemplateID')]
-      [string] $Id,
-
-      [Parameter(ParameterSetName = 'ByName', Mandatory = $true)]
-      [ProcessValidateAttribute()]
-      [ArgumentCompleter([ProcessTemplateCompleter])]
-      [string] $Name
+      [string] $Id
    )
    process {
-      if ($id) {
-         $queryString = @{ }
-
-         # Call the REST API
-         $resp = _callAPI -area 'process' -resource 'processes' -id $id `
-            -Version $(_getApiVersion Core) `
-            -QueryString $queryString -NoProject
-
-         $project = [VSTeamProcess]::new($resp)
-
-         Write-Output $project
+      # The REST API ignores Top and Skip but allows them to be specified & the function does the same. 
+      if ($PSBoundParameters['Top','Skip'] -gt 0) {
+         Write-Warning "You specified -Top $Top , -Skip $Skip These parameters are ignored and will be removed in future"
       }
-      elseif ($Name) {
-         # Lookup Process ID by Name
-         Get-VSTeamProcess | where-object { $_.name -eq $Name }
+      # Return either a single proces by ID or a list of processes
+      if ($id) {
+         # Call the REST API with an ID
+         $resp = _callAPI -NoProject -area 'work' -resource 'processes' -id $id  -Version $(_getApiVersion Core) 
+
+         if ($resp) {
+            return [VSTeamProcess]::new($resp)
+         }
       }
       else {
-         # Return list of processes
          try {
             # Call the REST API
-            $resp = _callAPI -area 'process' -resource 'processes' `
-               -Version $(_getApiVersion Core) -NoProject `
-               -QueryString @{
-               '$top'  = $top
-               '$skip' = $skip
-            }
-
-            $objs = @()
-            
-            foreach ($item in $resp.value) {
-               $objs += [VSTeamProcess]::new($item)
-            }
-
-            Write-Output $objs
+            $resp = _callAPI -NoProject -Area 'work' -resource 'processes' -Version (_getApiVersion core)  
          }
          catch {
             # I catch because using -ErrorAction Stop on the Invoke-RestMethod
@@ -58,6 +42,13 @@ function Get-VSTeamProcess {
             # This casuses the first error to terminate this execution.
             _handleException $_
          }
+         #we just fetched all the processes so let's update the cache. Also, cache the URLS for processes
+         [VSTeamProcessCache]::processes = $resp.value.name | Sort-Object
+         [VSTeamProcessCache]::timestamp = (Get-Date).Minute
+         $resp.value | ForEach-Object {
+            [VSTeamProcessCache]::urls[$_.name] =  (_getInstance) + "/_apis/work/processes/" + $_.TypeId
+            [VSTeamProcess]::new($_)
+         } | Where-Object {$_.name -like $Name} | Sort-Object -Property Name
       }
    }
 }